EZP-27646: Moved UserRegister namespace to more general User
