+++ lib/Form/Type/FieldType/UserAccountFieldType.php
@@ -5,8 +5,9 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace EzSystems\RepositoryForms\Form\Type\User;
+namespace EzSystems\RepositoryForms\Form\Type\FieldType;
 
+use EzSystems\RepositoryForms\Data\User\UserAccountFieldData;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
@@ -15,7 +16,7 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class UserAccountType extends AbstractType
+class UserAccountFieldType extends AbstractType
 {
     public function getName()
     {
@@ -24,25 +25,27 @@ class UserAccountType extends AbstractType
 
     public function getBlockPrefix()
     {
-        return 'ezuser';
+        return 'ezplatform_fieldtype_ezuser';
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $isUpdateForm = 'update' === $options['intent'];
+
         $builder
             ->add('username', TextType::class, [
-                'property_path' => 'username',
                 'label' => 'content.field_type.ezuser.username',
                 'required' => $options['required'],
+                'attr' => $isUpdateForm ? ['disabled' => 'disabled'] : [],
             ])
             ->add('password', RepeatedType::class, [
                 'type' => PasswordType::class,
-                'required' => true,
-                'property_path' => 'password',
+                'required' => $options['required'] && !$isUpdateForm,
                 'first_options' => ['label' => 'content.field_type.ezuser.password'],
                 'second_options' => ['label' => 'content.field_type.ezuser.password_confirm'],
             ])
             ->add('email', EmailType::class, [
+                'required' => $options['required'],
                 'label' => 'content.field_type.ezuser.email',
             ]);
     }
@@ -51,8 +54,10 @@ class UserAccountType extends AbstractType
     {
         $resolver
             ->setDefaults([
-                'data_class' => '\EzSystems\RepositoryForms\Data\User\UserAccountFieldData',
+                'data_class' => UserAccountFieldData::class,
                 'translation_domain' => 'ezrepoforms_content',
-            ]);
+            ])
+            ->setRequired(['intent'])
+            ->setAllowedValues('intent', ['register', 'create', 'update']);
     }
 }
