+++ lib/User/View/UserRegisterFormView.php
@@ -3,11 +3,10 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace EzSystems\RepositoryForms\UserRegister\View;
+namespace EzSystems\RepositoryForms\User\View;
 
 use eZ\Publish\Core\MVC\Symfony\View\BaseView;
-use eZ\Publish\Core\MVC\Symfony\View\View;
 
-class UserRegisterConfirmView extends BaseView implements View
+class UserRegisterFormView extends BaseView
 {
 }
